apps/backend/requirements.txt
fastapi
uvicorn
mysql-connector-python
openai
pydantic
sqlalchemy
langchain
langgraph
transformers
alembic
pymysql
